+++ tests/test_managers.py
@@ -27,26 +27,26 @@ class TestManagers(TestCase, WagtailTestUtils):
         sub_sub_page = TestPage(title="sub_sub_page")
         sub_page.add_child(instance=sub_sub_page)
 
-        self.assertEquals(TestPage.objects.count(), 3)
-        self.assertEquals(TestPage.objects_excluding_bins.count(), 3)
+        self.assertEqual(TestPage.objects.count(), 3)
+        self.assertEqual(TestPage.objects_excluding_bins.count(), 3)
 
         with self.register_hook("before_delete_page", trash_delete):
             delete_url = reverse("wagtailadmin_pages:delete", args=(sub_sub_page.id,))
             self.client.post(delete_url)
 
-        self.assertEquals(TestPage.objects.count(), 3)
-        self.assertEquals(TestPage.objects_excluding_bins.count(), 2)
+        self.assertEqual(TestPage.objects.count(), 3)
+        self.assertEqual(TestPage.objects_excluding_bins.count(), 2)
 
         with self.register_hook("register_admin_urls", urlconf_time):
             restore_url = reverse("wagtail_trash_restore", args=(sub_sub_page.id,))
             self.client.get(restore_url)
 
-        self.assertEquals(TestPage.objects.count(), 3)
-        self.assertEquals(TestPage.objects_excluding_bins.count(), 3)
+        self.assertEqual(TestPage.objects.count(), 3)
+        self.assertEqual(TestPage.objects_excluding_bins.count(), 3)
 
         with self.register_hook("before_delete_page", trash_delete):
             delete_url = reverse("wagtailadmin_pages:delete", args=(top_page.id,))
             self.client.post(delete_url)
 
-        self.assertEquals(TestPage.objects.count(), 3)
-        self.assertEquals(TestPage.objects_excluding_bins.count(), 0)
+        self.assertEqual(TestPage.objects.count(), 3)
+        self.assertEqual(TestPage.objects_excluding_bins.count(), 0)