+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 ### Fixed
 - Upgrade python version in example environment and resolve build issues
+- Replace deprecated assertEquals with assertEqual
 
 ### Removed
 - Drop EOL Python 3.7 support (@marteinn)